Test WinPrizeTest 테스트 케이스 수정

+++ src/test/java/lotto/domain/WinPrizeTest.java
@@ -9,7 +9,7 @@ public class WinPrizeTest {
     void 수익률_구하기_테스트() {
         long purchasedAmount = 20000;
         long totalPrice = Rank.FIRST.getPrize() + Rank.THIRD.getPrize();
-        long actual = totalPrice / purchasedAmount * 100;
+        double actual = (double) totalPrice / purchasedAmount * 100;
 
         WinPrize winPrize = new WinPrize();
         winPrize.addWinCount(Rank.FIRST);
